fix: tests for api
